+++ src/App/PropertyContainerPyImp.cpp
@@ -246,7 +246,7 @@ PyObject*  PropertyContainerPy::setPropertyStatus(PyObject *args)
             }
             status.set(it->second,value);
         }else if(item.isNumeric()) {
-            int v = Py::Int(item);
+            int v = Py::Long(item);
             if(v<0) {
                 value = false;
                 v = -v;
@@ -295,7 +295,7 @@ PyObject*  PropertyContainerPy::getPropertyStatus(PyObject *args)
                     }
                 }
                 if(!found)
-                    ret.append(Py::Int((long)i));
+                    ret.append(Py::Long((long)i));
             }
         }
     }
