+++ src/Gui/DocumentPyImp.cpp
@@ -414,7 +414,7 @@ Py::Object DocumentPy::getInEditInfo(void) const {
     if (!vp || !vp->getObject() || !vp->getObject()->getNameInDocument())
         return Py::None();
     return Py::TupleN(Py::Object(vp->getObject()->getPyObject(),true),
-            Py::String(subname),Py::String(subelement),Py::Int(mode));
+            Py::String(subname),Py::String(subelement),Py::Long(mode));
 }
 
 void DocumentPy::setInEditInfo(Py::Object arg) {
@@ -427,10 +427,10 @@ void DocumentPy::setInEditInfo(Py::Object arg) {
                 pyobj)->getViewProviderDocumentObjectPtr(),subname);
 }
 
-Py::Int DocumentPy::getEditMode(void) const {
+Py::Long DocumentPy::getEditMode(void) const {
     int mode = -1;
     getDocumentPtr()->getInEdit(0,0,&mode);
-    return Py::Int(mode);
+    return Py::Long(mode);
 }
 
 Py::Boolean DocumentPy::getTransacting() const {
