+++ src/Mod/Path/App/AreaPyImp.cpp
@@ -457,7 +457,7 @@ Py::List AreaPy::getShapes(void) const {
 	Area *area = getAreaPtr();
     const std::list<Area::Shape> &shapes = area->getChildren();
     for(auto &s : shapes)
-        ret.append(Py::TupleN(Part::shape2pyshape(s.shape),Py::Int(s.op)));
+        ret.append(Py::TupleN(Part::shape2pyshape(s.shape),Py::Long(s.op)));
     return ret;
 }
